oauth-refresh-daemon/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
apscheduler==3.10.4
google-auth==2.49.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.3.0
google-api-python-client==2.192.0
python-dotenv==1.0.1
